Feature created to update summoners in database.
There are verification of blank fields and if id exist in this aplication's database. Updating summoners can only be done by the user associated with it. Fixed the summoners removal and creation endpoints. User id was not being saved and removal of summoners can only be done by the user associated with it.